+++ src/components/pages/CardRegistrationPage.tsx
@@ -7,6 +7,7 @@ import CardForm from '../cardForm/CardForm';
 import { LeftArrowIcon } from '../../assets/icons';
 import type { CardItemInfo } from '../../types/Card';
 import BottomSheet from '../BottomSheet';
+import BankList from '../BankList';
 
 interface CardRegistrationPageProps {
   addCardItem: (cardItem: CardItemInfo) => void;
@@ -48,7 +49,7 @@ const CardRegistrationPage = ({ addCardItem }: CardRegistrationPageProps) => {
         onSubmitForm={handleSubmitForm}
         onChangeForm={handleChangeForm}
       />
-      <BottomSheet />
+      <BankList />
     </>
   );
 };
